+++ NEWS.md
@@ -2,4 +2,5 @@
 # austraits 1.1.2
 
 * Minor bug fix in `extract_trait`
-* `site` or `site_name` has been renamed to `location` and `location_name`
+* `site` & `site_name` renamed to `location` & `location_name` within functions
+* `austraits$sites` renamed to `austraits$locations`